+++ engine/language-server/src/main/scala/org/enso/languageserver/LanguageServer.scala
@@ -1,84 +1,23 @@
 package org.enso.languageserver
 import java.io.File
-import java.util.UUID
 
 import akka.actor.{Actor, ActorLogging, ActorRef, Stash}
-import org.enso.languageserver.data.{CanEdit, CapabilityRegistration}
+import org.enso.languageserver.data.{CanEdit, CapabilityRegistration, Client, Config, Env}
 
 object LanguageProtocol {
-  type ClientId = UUID
   case class Initialize(config: Config)
-  case class Connect(clientId: ClientId, clientActor: ActorRef)
-  case class Disconnect(clientId: ClientId)
-
-  case class Config(contentRoots: List[File], languagePath: List[File])
-
-  case class Client(
-    id: ClientId,
-    actor: ActorRef,
-    capabilities: List[CapabilityRegistration]
-  )
-
-  case class Env(clients: List[Client]) {
-    def addClient(client: Client): Env = {
-      copy(clients = client :: clients)
-    }
-
-    def removeClient(clientId: ClientId): Env =
-      copy(clients = clients.filter(_.id != clientId))
-
-    def removeCapabilitiesBy(
-      predicate: CapabilityRegistration => Boolean
-    ): (Env, List[(Client, List[CapabilityRegistration])]) = {
-      val newClients = clients.map { client =>
-        val (removedCapabilities, retainedCapabilities) =
-          client.capabilities.partition(predicate)
-        val newClient = client.copy(capabilities = retainedCapabilities)
-        (newClient, removedCapabilities)
-      }
-      (copy(clients = newClients.map(_._1)), newClients)
-    }
-
-    def modifyClient(
-      clientId: ClientId,
-      modification: Client => Client
-    ): Env = {
-      val newClients = clients.map { client =>
-        if (client.id == clientId) {
-          modification(client)
-        } else {
-          client
-        }
-      }
-      copy(clients = newClients)
-    }
-
-    def grantCapability(
-      clientId: ClientId,
-      registration: CapabilityRegistration
-    ): Env =
-      modifyClient(clientId, { client =>
-        client.copy(capabilities = registration :: client.capabilities)
-      })
-
-    def releaseCapability(clientId: ClientId, capabilityId: UUID): Env =
-      modifyClient(clientId, { client =>
-        client.copy(
-          capabilities = client.capabilities.filter(_.id != capabilityId)
-        )
-      })
-  }
-
-  object Env {
-    def empty: Env = Env(List())
-  }
+  case class Connect(clientId: Client.Id, clientActor: ActorRef)
+  case class Disconnect(clientId: Client.Id)
 
   case class AcquireCapability(
-    clientId: UUID,
+    clientId: Client.Id,
     registration: CapabilityRegistration
   )
-  case class ReleaseCapability(clientId: ClientId, capabilityId: UUID)
-  case class ForceReleaseCapability(capabilityId: UUID)
+  case class ReleaseCapability(
+    clientId: Client.Id,
+    capabilityId: CapabilityRegistration.Id
+  )
+  case class ForceReleaseCapability(capabilityId: CapabilityRegistration.Id)
   case class GrantCapability(registration: CapabilityRegistration)
 }
 
@@ -97,11 +36,13 @@ class Server extends Actor with Stash with ActorLogging {
     case Connect(clientId, actor) =>
       log.debug("Client connected [{}].", clientId)
       context.become(
-        initialized(config, env.addClient(Client(clientId, actor, List())))
+        initialized(config, env.addClient(Client(clientId, actor)))
       )
+
     case Disconnect(clientId) =>
       log.debug("Client disconnected [{}].", clientId)
       context.become(initialized(config, env.removeClient(clientId)))
+
     case AcquireCapability(
         clientId,
         reg @ CapabilityRegistration(_, capability: CanEdit)
@@ -118,6 +59,7 @@ class Server extends Actor with Stash with ActorLogging {
       }
       val newEnv = envWithoutCapability.grantCapability(clientId, reg)
       context.become(initialized(config, newEnv))
+
     case ReleaseCapability(clientId, capabilityId) =>
       context.become(
         initialized(config, env.releaseCapability(clientId, capabilityId))
