+++ engine/language-server/src/main/scala/org/enso/languageserver/data/Client.scala
@@ -1,0 +1,15 @@
+package org.enso.languageserver.data
+import java.util.UUID
+
+import akka.actor.ActorRef
+
+case class Client(
+  id: Client.Id,
+  actor: ActorRef,
+  capabilities: List[CapabilityRegistration]
+)
+
+object Client {
+  type Id = UUID
+  def apply(id: Id, actor: ActorRef): Client = Client(id, actor, List())
+}
